# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApoTrack API",
        "description": "Dokumentasi API resmi untuk platform ApoTrack — sistem manajemen apotek digital. API ini digunakan oleh aplikasi mobile Flutter untuk Customer dan Staff Apotek.",
        "contact": {
            "name": "Tim Pengembang ApoTrack",
            "email": "dev@apotrack.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Aktif (Local Development)"
        }
    ],
    "paths": {
...
        "/api/orders/history": {
            "get": {
                "tags": [
                    "Orders (Customer)"
                ],
                "summary": "Riwayat pesanan selesai/batal Customer",
                "description": "Mengambil daftar pesanan milik pengguna yang berstatus COMPLETED atau CANCELLED.",
                "operationId": "getCustomerOrderHistory",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman (default: 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat pesanan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Riwayat pesanan berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders (Customer)"
                ],
                "summary": "Detail pesanan Customer",
                "description": "Mengambil detail pesanan lengkap beserta item obat, status log, dan informasi resep jika ada.",
                "operationId": "getCustomerOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pesanan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail pesanan berhasil diambil"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Order]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}/tracking": {
            "get": {
                "tags": [
                    "Orders (Customer)"
                ],
                "summary": "Pelacakan pengiriman kurir (Biteship)",
                "description": "Mengambil riwayat pelacakan kurir secara real-time dari Biteship untuk pesanan dengan layanan DELIVERY.",
                "operationId": "getOrderTracking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pelacakan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data pelacakan kurir berhasil diambil"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pelacakan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Informasi pengiriman tidak ditemukan untuk pesanan ini."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum Anda di sini. Token didapatkan dari endpoint POST /api/auth/login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}